requirements.txt
groq
python-dotenv
playwright
websockets
pandas
asyncpg
redis
colorama
